requirements.txt
ccxt==1.91.75
python-binance==1.0.12
python-dotenv==0.20.0
PyQt5==5.15.7
plotly==5.9.0
pandas==1.4.2
SQLAlchemy==1.4.40
PyQt5==5.15.7
PyQtChart==5.15.6
PyQtChart-Qt5==5.15.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1